[IMP] web_m2x_options: Disable hardcoded "Search More..." limit when no search value is set.
To improve `name_get` performance, there is an hardcoded limit.
But when no search value is set, this name_get query is useless since only the domain will limit the result.
So, when the field input is empty and when the user clicks on "Search More...", we let the Form doing its 'web/dataset/search_read' query without locking the domain on `'[id', 'in', (...)]`.